Creates and updates items in a list
Created a simple module and routes for creating and updating items
within a single list. PUT and POST requests are used to update and
create items respectively.

I chose to go with callbacks over promises to not introduce any
additional talking points during the presentation. If that's a
non-issue, I can convert these over to promises.